backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.12
pydantic-settings==2.13.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-api-python-client==2.193.0
pywebpush==2.3.0
py-vapid==1.9.4
slowapi==0.1.9
aiofiles==23.2.1
httpx==0.28.1
